changes required to use telemed schedule instead of hardcoded pt. 2
