+++ packages/utils/lib/fhir/location.ts
@@ -1,7 +1,26 @@
 import Oystehr from '@oystehr/sdk';
-import { Location, Resource, Schedule } from 'fhir/r4b';
-import { TelemedLocation } from '../types';
-import { PUBLIC_EXTENSION_BASE_URL } from './constants';
+import {
+  Encounter,
+  HealthcareService,
+  Location,
+  LocationHoursOfOperation,
+  Practitioner,
+  Resource,
+  Schedule,
+} from 'fhir/r4b';
+import { DateTime } from 'luxon';
+import { getClosingTime, getOpeningTime, isClosureOverride } from '../helpers';
+import {
+  AvailableLocationInformation,
+  HOURS_OF_OPERATION_FORMAT,
+  OVERRIDE_DATE_FORMAT,
+  ScheduleListItem,
+  ScheduleType,
+  TelemedLocation,
+  TIMEZONES,
+} from '../types';
+import { DOW, getScheduleDetails, getTimezone } from '../utils';
+import { PUBLIC_EXTENSION_BASE_URL, SLUG_SYSTEM } from './constants';
 
 export const isLocationFacilityGroup = (location: Location): boolean => {
   return Boolean(
@@ -78,15 +97,14 @@ export async function getTelemedLocations(oystehr: Oystehr): Promise<TelemedLoca
     (location) => location.resourceType === 'Location' && isLocationVirtual(location)
   ) as Location[];
 
-  const locationToScheduleMap = new Map<string, string | undefined>();
+  const locationToScheduleMap = new Map<string, Schedule | undefined>();
   resources.forEach((res) => {
     const isSchedule = res.resourceType === 'Schedule';
     if (isSchedule) {
-      const scheduleId = res.id;
       const actor = res.actor?.find((a) => a.reference?.startsWith('Location'))?.reference;
       if (actor) {
         const locationId = actor.split('/')[1];
-        locationToScheduleMap.set(locationId, scheduleId);
+        locationToScheduleMap.set(locationId, res);
       }
     }
   });
@@ -164,4 +182,174 @@ export const defaultLocation: Location = {
       daysOfWeek: ['sun'],
     },
   ],
+};
+
+// todo 1.8: this needs to take a schedule (or be async and go get a schedule), have a better name
+// also check that this data is truly needed everywhere it is used
+export function getLocationInformation(
+  oystehr: Oystehr,
+  scheduleResource: Location | Practitioner | HealthcareService,
+  currentDate: DateTime = DateTime.now()
+): AvailableLocationInformation {
+  const slug = scheduleResource.identifier?.find((identifierTemp) => identifierTemp.system === SLUG_SYSTEM)?.value;
+  const timezone = getTimezone(scheduleResource);
+
+  const schedule = getScheduleDetails(scheduleResource);
+  const scheduleOverrides = schedule?.scheduleOverrides || {};
+
+  let scheduleType: ScheduleType;
+  switch (scheduleResource?.resourceType) {
+    case 'Location':
+      scheduleType = ScheduleType['location'];
+      break;
+    case 'HealthcareService':
+      scheduleType = ScheduleType['group'];
+      break;
+    case 'Practitioner':
+      scheduleType = ScheduleType['provider'];
+      break;
+  }
+
+  // Modify hours of operation returned based on schedule overrides
+  let hoursOfOperation: LocationHoursOfOperation[] | undefined = undefined;
+  if (scheduleResource.resourceType === 'Location') {
+    hoursOfOperation = scheduleResource.hoursOfOperation;
+    currentDate = currentDate.setZone(timezone);
+    const overrideDate = Object.keys(scheduleOverrides).find((date) => {
+      return currentDate.toFormat(OVERRIDE_DATE_FORMAT) === date;
+    });
+    if (overrideDate) {
+      const dayOfWeek = currentDate.toFormat('EEE').toLowerCase();
+      const override = scheduleOverrides[overrideDate];
+      const dayIndex = hoursOfOperation?.findIndex((hour) => (hour.daysOfWeek as string[])?.includes(dayOfWeek));
+      if (hoursOfOperation && typeof dayIndex !== 'undefined' && dayIndex >= 0) {
+        hoursOfOperation[dayIndex].openingTime = DateTime.fromFormat(override.open.toString(), 'h')
+          .set({
+            year: currentDate.year,
+            month: currentDate.month,
+            day: currentDate.day,
+          })
+          .toFormat(HOURS_OF_OPERATION_FORMAT);
+        hoursOfOperation[dayIndex].closingTime = DateTime.fromFormat(override.close.toString(), 'h')
+          .set({
+            year: currentDate.year,
+            month: currentDate.month,
+            day: currentDate.day,
+          })
+          .toFormat(HOURS_OF_OPERATION_FORMAT);
+      }
+    }
+  }
+
+  return {
+    id: scheduleResource.id,
+    slug: slug,
+    name: getName(oystehr, scheduleResource),
+    description: undefined,
+    address: undefined,
+    telecom: scheduleResource.telecom,
+    hoursOfOperation: hoursOfOperation,
+    timezone: timezone,
+    closures: schedule?.closures ?? [],
+    otherOffices: [], // todo
+    scheduleType,
+  };
+}
+
+function getName(oystehrClient: Oystehr, item: Location | Practitioner | HealthcareService): string {
+  if (!item.name) {
+    return 'Unknown';
+  }
+  if (item.resourceType === 'Location') {
+    return item.name;
+  }
+  if (item.resourceType === 'HealthcareService') {
+    return item.name;
+  }
+  return oystehrClient.fhir.formatHumanName(item.name[0]);
+}
+
+export function getEncounterStatusHistoryIdx(encounter: Encounter, status: string): number {
+  if (encounter.statusHistory) {
+    return encounter.statusHistory.findIndex((history) => history.status === status && !history.period.end);
+  } else {
+    throw new Error('Encounter status history not found');
+  }
+}
+
+export interface CheckOfficeOpenOutput {
+  officeOpen: boolean;
+  walkinOpen: boolean;
+  prebookStillOpenForToday: boolean;
+  officeHasClosureOverrideToday: boolean;
+  officeHasClosureOverrideTomorrow: boolean;
+}
+
+export function checkOfficeOpen(location: AvailableLocationInformation): CheckOfficeOpenOutput {
+  const timeNow = DateTime.now().setZone(location.timezone);
+  const tomorrowDate = timeNow.plus({ days: 1 });
+  const tomorrowOpeningTime = getOpeningTime(location.hoursOfOperation ?? [], location.timezone ?? '', tomorrowDate);
+
+  const officeHasClosureOverrideToday = isClosureOverride(location.closures ?? [], location.timezone ?? '', timeNow);
+  const officeHasClosureOverrideTomorrow =
+    isClosureOverride(location.closures ?? [], location.timezone ?? '', tomorrowDate) &&
+    tomorrowOpeningTime !== undefined;
+
+  const todayOpeningTime = getOpeningTime(location.hoursOfOperation ?? [], location.timezone ?? '', timeNow);
+  const todayClosingTime = getClosingTime(location.hoursOfOperation ?? [], location.timezone ?? '', timeNow);
+  const prebookStillOpenForToday =
+    todayOpeningTime !== undefined && (todayClosingTime === undefined || todayClosingTime > timeNow.plus({ hours: 1 }));
+
+  const officeOpen =
+    todayOpeningTime !== undefined &&
+    todayOpeningTime <= timeNow &&
+    (todayClosingTime === undefined || todayClosingTime > timeNow) &&
+    !officeHasClosureOverrideToday;
+
+  return {
+    officeOpen,
+    walkinOpen: officeOpen,
+    prebookStillOpenForToday,
+    officeHasClosureOverrideToday,
+    officeHasClosureOverrideTomorrow,
+  };
+}
+
+export const getHoursOfOperationForToday = (item: Schedule): ScheduleListItem['todayHoursISO'] => {
+  const tz = getTimezone(item) ?? TIMEZONES[0];
+  const dayOfWeek = DateTime.now().setZone(tz).toLocaleString({ weekday: 'long' }).toLowerCase();
+
+  const scheduleTemp = getScheduleDetails(item);
+  if (!scheduleTemp) {
+    return undefined;
+  }
+  const scheduleDays = scheduleTemp.schedule;
+  const scheduleDay = scheduleDays[dayOfWeek as DOW];
+  let open: number = scheduleDay.open;
+  let close: number = scheduleDay.close;
+  const scheduleOverrides = scheduleTemp.scheduleOverrides;
+  if (scheduleTemp.scheduleOverrides) {
+    for (const dateKey in scheduleOverrides) {
+      if (Object.hasOwnProperty.call(scheduleOverrides, dateKey)) {
+        const date = DateTime.fromFormat(dateKey, OVERRIDE_DATE_FORMAT).setZone(tz).toISODate();
+        const todayDate = DateTime.now().setZone(tz).toISODate();
+        if (date === todayDate) {
+          open = scheduleOverrides[dateKey].open;
+          close = scheduleOverrides[dateKey].close;
+        }
+      }
+    }
+  }
+  if (open !== undefined && close !== undefined) {
+    const openTime = DateTime.now().setZone(tz).startOf('day').plus({ hours: open }).toISO();
+    const closeTime = DateTime.now().setZone(tz).startOf('day').plus({ hours: close }).toISO();
+    if (!openTime || !closeTime) {
+      return undefined;
+    }
+    return {
+      open: openTime,
+      close: closeTime,
+    };
+  }
+  return undefined;
 };
